[Edited]: data type ของทุก id  ใน table จาก  int -> bigint

+++ tearup/init.sql
@@ -11,7 +11,7 @@ INSERT INTO user VALUE (1,"sckshuhari");
 
 
 CREATE TABLE products (
-    id int AUTO_INCREMENT,
+    id BIGINT AUTO_INCREMENT,
     product_name varchar(255),
     product_brand varchar(255),
     quantity int,
@@ -25,7 +25,7 @@ CREATE TABLE products (
 INSERT INTO products (id,product_name,product_brand,quantity,product_price,image_url) VALUE (2,"43 Piece dinner Set","CoolKidz",10,12.95,"/43_Piece_dinner_Set.png");
 
 CREATE TABLE orders (
-    id int AUTO_INCREMENT,
+    id BIGINT AUTO_INCREMENT,
     total_price double,
     transaction_id varchar(255) DEFAULT '',
     completed smallint(1) DEFAULT 0,
@@ -36,15 +36,15 @@ CREATE TABLE orders (
 );
 
 CREATE TABLE order_product (
-    order_id int,
-    product_id int,
+    order_id BIGINT,
+    product_id BIGINT,
     quantity int,
     product_price double
 );
 
 CREATE TABLE shipping (
     id int AUTO_INCREMENT,
-    order_id int,
+    order_id BIGINT,
     address varchar(255),
     sub_district varchar(255),
     district varchar(255),
